+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   not the entire body and doc comment.
 - update `unnecessary_getters_setters` to allow otherwise "unnecessary" getters
   and setters with annotations.
+- update `missing_whitespace_between_adjacent_strings` to allow String
+  interpolations at the beginning and end of String literals.
 
 # 1.6.1